requirements.txt
requests
rich
prompt_toolkit
openai
pyyaml
click
python-dotenv
pathlib
datetime
